Babel: allow optional chaining (replacement for IDX)
This new way of writing optional chain is much more nicer and easier to use. Other benefit is that we can check the usage more precisely - if you use it and it's not necessary then Flow will throw an error like this:

> This use of optional chaining (?.) is unnecessary because XYZ [1] cannot be nullish or because an earlier ?. will short-circuit the nullish case. (unnecessary-optional-chain)

Please note: I didn't remove all the IDX occurrences. We will migrate it over the time. We should not allow IDX in new MRs though.