Implement workspace dependent testing
Only not enabled for CI yet. This is how it works: it checks whether something changed in the workspace and if so then it tests the workspace itself as well as all dependencies (workspace can be dependent on something or it is itself a dependency). For example: when you change something in `vault2env` then it tests only this package because we do not use it anywhere. But when you change for example `global-id` then this workspace is tested as well as basically the whole application. As a nice consequence these tests are not being run for README changes for example. They are still being deployed though even though it's not necessary (that a different problem to solve somewhere in the future).

Closes: https://gitlab.skypicker.com/graphql/graphql/issues/386