Turn core application into workspaces
Benefits are not very visible at this moment but this is very useful for dependent testing among all the workspaces. I am preparing a script which will be able to find all relevant dependencies across our workspaces and run only the necessary subset of tests. Thanks to this change we can skip testing of the whole core if not needed (because we touched only some unrelated package for example).

Also, I found one bug related to workspaces so it's handy to keep everything the same so we can leverage from the ecosystem we are building here.