Enable dependent testing in CI
I made a mistake and it is already partially enabled. I think there is no reason to wait for it and the best way how to prove this concept is to just use it. So, here we  go... :) And just to make sure everything is gonna be fine I am still running all the tests before deploy (does affect only deployment stage).