Improve SmartFAQ query tests
This basically replaces many many lines of copy-paste for something more automatic I learned while studying Relay codebase. It should be enough for these tests just to write the query and save the output - that's it. We should connect it to something like Skymock (investigating further) and this way we could remove even the `datasetMap` completely. Great benefit of this approach is that you can use GraphQL fragments across your tests (see for example `allFAQCategories.graphql`). We should also educate people that this is not the right way how to test resolvers (via different query variants) - these tests are too highlevel for resolvers.