Remove usages of 'toGlobalId'
I am not 100% sure about the solution but I am pretty sure about removing this function. I think it's was a bad pattern because people had to copy the name of the GraphQL object which made it incredibly fragile (see the corrected snapshots) and weird. It is most visible on the 'arrayconnection' hack. How did everyone know it should be called exactly like this? It's just an internal implementation detail... :)

Closes https://gitlab.skypicker.com/graphql/graphql/issues/366