Automator: implement replacement tags autodiscovery
Automator is now able to automatically find tags to replace in our monorepo. At this moment it's tightly coupled with the "update NPM packages info" task since that's the only task we have. But the idea can be later expanded to every task (and file type). Important is that we don't have to manually set where the automator acts - just place the replacement tags somewhere and it will be replaced automatically.