Install Node.js dependencies before all CI jobs
There is no very obvious benefit at this moment except this is how it should be done. The point is to progressively test and build the application throughout the stages and this is exactly it: it installs the Node.js dependencies only once and later stages just use them. This means we don't have to mess up with artifacts and cache in later stages. I also added JUnit reporter which creates a nice results in GitLab CI - just to try it.