Eslint: enable warnings for 'require-await' rule
Seems like this is quite common mistake. I had to upgrade our eslint runner to support warnings for successful messages. Also, I didn't fix these warnings everywhere - only the obvious cases. This is because we abused async/await behavior in some cases and I have to be careful there to fix it properly.