GraphQL Skymock: use stable JSON stringify
All snapshots are not going to be stored in a stable form (more or less because our APIs responses are sometimes very unstable). This is not going to reduce the number of changes submitted by GraphQL Skymock but I hope it will solve our flaky tests when snapshots are different only because of the order.