Babel: transpile 'warning' function from '@kiwicom/js'
This allows us to completely remove warning functions from the code when bundling the application (with dead code elimination). Something similar will be done for invariant but it's not identical (invariant must stay in the code). Please note: warning will not be called in production even if you do not use this Babel plugin. However, it will not be removed from the final bundle.

Related issue: https://gitlab.skypicker.com/incubator/universe/issues/432