GraphQL Skymock: add initial support for POST requests
Ancillaries are the first API with POST requests that is being updated automatically. It's because it doesn't need any auth. I have to focus on a bit of refactoring in the next iterations because it's getting a bit complicated to develop it - there are not very clear boundaries and rules when using the meta file.