NPM: stop publishing '@kiwicom/logzio'
We originally did this only because of BLKN but they never used it and they eventually lost interest. So I think it's gonna be better to keep it internal for now so we don't have to deal with the releasing. We could eventually rename it to `@kiwicom/graphql-monitoring` and connect it to Datadog. I will deprecate this package on NPM once this is merged.