Automator: remove 'yarnUpgrade' script
It's not good. Especially not with our new offline mirror. It would be great to replace it with some ready-made solution but I am afraid it's not gonna be good as well because monorepo env is very specific. For instance, there are some dependencies we don't want to upgrade (typically Expo has locked low React version and exotic RN version).