Add Monorepo Scanner
Purpose of this Scanner is to run checks which are outside of Tests Runner scope.

Typical problem we currently have: package.json version is changed in one workspace and all the tests (and their related tests) are executed correctly. However, test checking changelog is not executed because there is no clear relation between this test and the package.json. This is problem for any other check outside of workspaces. This Scanner solves this problem.

My goal is to help developers to follow the monorepo rules correctly. For example, I added one check which checks that Next.js application (Babel) is configured correctly - that would be also quite complicated to verify using only Jest and Tests Runner.