GraphQL: init Datadog in index file
Previously, we had to do it `Application` file to manually instrument GraphQL traces. It's not necessary now so I'd like to init it as soon as possible. There is not much more to do from our side to make it work. Unfortunately, it's not clear if this is gonna work or not: `dd-trace` is pure magic. This also fixes `Dockerfile-new` build.