Shipit: add patch fixture for LFS
We are currently not using it but I want to be sure it's fine in advance. Shipit does not support sync of these LFS files, however, we are about to use LFS only for Yarn offline mirror (maybe) and these files are not being synced to GitHub anyway so it's not needed (yet).