Docs: add basic info about GCP
This is still undiscovered area in Universe. I am starting first kickoff with this page and I'll update it as we go. It's still changing a lot and it's not very stable. Ultimate goal here is to have really simple way how to start things in GCP for any project in Universe.