Eslint: allow non null values directly in a list
Valid and invalid values by list type definition possibilities (using SDL for simplicity):

1. `[String]`
    - valid values: `null`, `[]`, `['foo', 'bar']`, `[null, null]`
2. `[String]!`
    - valid values: `[]`, `['foo', 'bar']`, `[null, null]`
    - invalid values: `null`
3. `[String!]`
    - valid values: `null`, `[]`, `['foo', 'bar']`
    - invalid values: `[null, null]`
4. `[String!]!`
    - valid values: `[]`, `['foo', 'bar']`
    - invalid values: `null`, `[null, null]`

The current rule requires 1st possibility (above) only everywhere, however, allows all possibilities in input types and `args`.

It does not make much sense to have nullable values in a list usually so we want them to be not null (the 3rd and 4th possibility above).

It can also be updated (in a future version?) to require list values not to be null always. At least for input types.