Flowconfig: remove rollouts
I tried and I don't like it. The biggest deal breaker for me is that the rollouts are being enabled per machine and not per fail. Therefore you can rollout the whole features (which was the original intention) but you cannot rollout them file by file. That's not friendly. I will ask Flow devs if there is a chance to improve this behavior.