Flow: add simple 'strictify' script
This is a pretty naive script which automatically adds strict annotations where possible. It does so by simply iterating all the files and trying to change the Flow annotations to strict ones (and reverting if not possible). Therefore you might want to run it multiple times to achieve the best effect.