Relay compiler: remove experimental custom validators
Validators were removed from Relay Compiler in 7.0.0. The right strategy would be to rewrite them to transforms, however, transforms use IR representation of AST which is not compatible with GraphQL.js implementation. Therefore, it is not so straightforward. I might return them back eventually but priority is to upgrade to 7.0.0.