GRS: improve scoring of fragment spreads
Previously, we calculated the score for fragments separately and ignored the spread completely. This is, however, not accurate since the spread can be in the list which negatively affects the final score. This new approach calculates the fragment in a place of the spread which give much more precise (and higher) score as a result.

Please note: there are still some caveats which I didn't cover. First of all, there is no additional performance optimization (we can calculate the score only once per fragment, not every time). There is also no fragment or field deduplication which leads to slightly inaccurate results in some cases.