GraphQL ID: relax `__isTypeOf` input to accept anything
We have some cases where this function crashes weirdly like in the following example when people don't use it correctly (not sending opaque ID as an input):

```
Caused by: TypeError: Cannot read property 'replace' of undefined
    at decode (/app/node_modules/@adeira/graphql-global-id/src/Encoder.js:34:24)
    at Object.__isTypeOf (/app/node_modules/@adeira/graphql-global-id/src/GlobalID.js:45:45)
    at GraphQLObjectType.isTypeOf (/app/src/http/graphql/user/types/User.ts:62:37)
    at completeObjectValue (/app/node_modules/graphql/execution/execute.js:686:31)
    at completeValue (/app/node_modules/graphql/execution/execute.js:591:12)
    at /app/node_modules/graphql/execution/execute.js:492:16
    at runMicrotasks (<anonymous>)
    at processTicksAndRejections (internal/process/task_queues.js:97:5)
    at async Promise.all (index 0)
    at middleware (/app/node_modules/koa-graphql/dist/index.js:117:16)
```

It's convenient to allow `any` as an input to be able to send there some false values (like null or undefined). One example being:

```
isTypeOf: ({ nodeId }): boolean => __isTypeOf(TypeName, nodeId),
```

Where `nodeId` can be `undefined` (which is hard to know in advance). Moreover, it makes sense to allow anything because we are asking whether "is type of" and not whether "is opaque ID type of" (even though that's arguable).