eslint-runner: Improve lint all logic
- Only lint all if dependency is removed or eslintConfig field is changed.

The motivation for this improvement is that it is unnecessary to lint
all files when doing changes that do not relate to eslint.
Like upgrading a dependency or adding a script.
What is relevant for eslint is removal of dependency. You need
to check that the dependency is not used in the code.
If package.json contains eslintConfig field that changed,
we also need to lint all files