Relay runtime: add "lazy" and "debug" logger
I've seen something similar in one FB talk which make me realize that we are logging unnecessary garbage. So I renamed the current logger to "eager" logger and created a new "lazy" one. The difference is, that this logger waits for the response and prints all the information at once (instead of streaming them into console as they arrive). This is a first iteration but eventually I'd like to use the lazy logger as a default one inside our Relay package.