SX: remove 'isUnitlessNumber' from the Flow type generator
The problem is, that properties inside 'isUnitlessNumber' list are describing unitless numbers but they do not say anything about the expected property type. For example:

```
zIndex: 10       # is unitless, good, but type number is not enough
zIndex: 'auto'   # can be also a string
```

Therefore, the resulting Flow type should not be affected by the "unitlessness" of the property.

This change reverts the generator back to always generate `string | number` by default which will be more precise and accurate. All these types can be overwritten in `CustomPropertyTypes` if needed.