SX: improve generated Flow types using 'mdn-data'
This makes some types less accurate because I removed `CustomPropertyTypes` but at the same time it makes many properties more accurate because we actually have a source for these props. We should keep iterating on the generator and effectively make the types more accurate (instead of writing the types manually).