SX: Render media queries as last on the page
In CSS the rule that is set last, is the one that will be executed. That means rules in media queries would never be used if they are not at the end.

Here the red always wins:
```
@media (max-width: 640px) {
  color: blue;
}
color: red;
```

Here the color is blue on a screen smaller than 640px:
```
color: red;
@media (max-width: 640px) {
  color: blue;
}
```