SX: refactor StyleCollector
The goal of this commit is to cleanup the style collector and prepare it for runtime styles. Specifically, I didn't like that the collector is mingling StyleNode and StyleCollector itself together. So I separated these concerns into individual nodes. These nodes also know how to print themselves so we don't need the CSS renderers.

I also modified the interface of the collector to return style buffer and hash registry instead of mutating the input. This simplified the usage even further.