SX: expand shorthand properties
Purpose of this change is to expand shorthand CSS properties which could cause conflics when defined together. Imagine the following example:

```
const styles = sx.create({
  primary: { marginTop: '10px' },
  button: { margin: 0 },
});
```

It could generate the following CSS:

```
.c0 { margin-top: 10px }
.c1 { margin: 0px }
```

That would be a tricky situation because CSS specificity depends on the CSS definition order and therefore the top margin would be always overwritten when used together. Instead, we expand the shorthand properties into something like this:

```
.c0 { margin-top: 10px }   << primary
.c1 { margin-top: 0px }    << button
.c2 { margin-right: 0px }
.c3 { margin-bottom: 0px }
.c4 { margin-left: 0px }
```

This way SX is not dependent on the CSS rules insertion order but rather on the markup.

Inspired by this article: https://sebastienlorber.com/atomic-css-in-js (and React Native Web)