Open-source new 'eslint-plugin-sx'
This plugin has only one rule at this moment: `sx/no-unused-stylesheets`. This rule is trying to capture cases when the defined stylesheet is not used anywhere in the application (manual dead code elimination) or when used stylesheet was not defined.

It also validates the SX usage a little bit which should probably be moved into some other rule (to be done later).