sx: Add keyframes support
This commit adds a keyframes map to StyleCollector and exports a new
keyframes function. The keyframes function parses the animation object
and generates a hashed name based on the input. Then it passes
the generated style and hash to the StyleCollector. The StyleCollector
reports back if the animation already exists. If it does not exists
and we are in the browser, the function appends the animation to
the SX stylesheet definitions.