Extract Flow types for Eslint plugins/rules
I need to reuse Flow types for the new `eslint-plugin-sx` so I made them reusable via `@adeira/flow-types-eslint` package. I changed the types a little bit to make them more accurate which resulted in new errors. The idea is to iterate on these types as needed (or replace them if we find something better).