SXT: Use Tailwind setup
This is a fundamental rework of SX Tailwind. Previously the styles definitions for SX were generated from a Tailwind CSS file, now it's generated from the same data the Tailwind itself  is generated.

It opens the way for a full support of customisation and simpler upgrade when a new version of Tailwind is released.