Eslint SX: improve error nodes location
Previously, we were reporting the errors with 'Program' node which resulted in the whole file being marked as errored. That's a bad practice since you need to go and search for the actual error location. This commit improves the behavior by reporting the errors with relevant nodes. This results in much better error location reporting.