Improve Flow server starts when saved state fails
Previously, we were halting the Flow server start when saved state could not be loaded. That was basically never the case since our setup was quite simple. However, now, we have 2 variants (iOS and Android) and these 2 saved states are not compatible (causing Flow server to start from scratch).

That's an issue because this newly created server was not lazy. So I changed the behavior a little bit:

- Flow server always starts with the lazy mode (even if it failed to load the saved state) - removed `--saved-state-no-fallback`
- the consequent commands always expect the server to be started (and never start it itself incorrectly like before) - added `--no-auto-start`