CI: setup Rust cross compilation (Apple, Linux)
To make sure the code works for both Apple and Linux platforms. CI for each platform is slightly different: for Apple we run only tests to make sure it works for local development whereas on Linux we run Clippy, tests and release build (the assumption is that production uses Linux). This also saves some CI time on Macos machine.