Relay runtime: fix broken Flow types
Somehow, we merged a PR with Relay 10.1.1 without catching the Flow errors on CI. Here is what _I think_ happened:

- the PR changed only `package.json` files which we correctly written into `.flow.saved_state_file_changes`
- we started the server in a lazy FS mode which initialized with 0 changed files (also correct)
- we called `flow force-recheck` for these `package.json` files which however doesn't refocus well enough and doesn't have any effect on the typechecking -> still 0 changed files
- Flow said "OK" even though there were errors

This workflow usually works for JS files but it doesn't work when no JS files changed (in the previous scenario only JSON files changed).

I created a TODO for myself to improve the workflow, possibly to remove the lazy mode completely (or run Flow properly on CI without shortcuts).