YCR: remove 'sections' collection
I came to a realization that `sections` collection is essentially an edge in the graph database and therefore should be completely removed and replaced with the actual edge.

Apart from that I also improved the AQL queries to reflect recommended best practices and used named graphs instead of anonymous graphs.