YC: implement sessions management (auth/deauth)
This commit implements `authorize_mobile` and `deauthorize_mobile` mutations as well as `whoami` query and request verification. This way, we can login on mobile via Google Sign-In, exchange the token on server for our session token and use it for the following requests. Eventually, we can easily implement more providers.