Add '@adeira-github-bot' to CODEOWNERS
Seems like there is currently no way of having a single code owner and be able to do self-approve & merge. Adeira GitHub Bot could help with this since it can behave as a normal user.

I am still looking for a way how to automate it but before I do so, I want to verify this is a good strategy.