YCR: make GraphQL schema a thin layer
Previously, GraphQL schema lived under SDUI which doesn't make sense anymore. This commit moves the files on a "server" level and makes the GraphQL query as thin as possible. It still has to be registered all in one file, unfortunately, but this might improve in future Juniper versions (there are several issues talking about this and about God Context objects).