Rust CI: remove paths conditional
Unfortunately, due to GitHub CI limitations, we cannot limit the workflow for certain paths and at the same time make the CI jobs required (for branch protection). I'd like to remove the paths since it's more important to have reliably green CI.

Possible improvements in the future: we can use special action instead which runs only on certain paths. That would workaround this issue nicely. However, it's probably not needed now since we have a Kodiak and merging times are not a big deal.