YCR: hash session tokens
Last missing piece for the session management. Session tokens are essentially long-time passwords and we should store them securely as well. This commits tries to find a balance between security and performance since we have to work with the hashes per API request (PBKDF2 with many iterations is too slow and unnecessary in our case).