KOCHKA: revert `LanguageTag` test
This is an attempt to partially revert `LanguageTag` test which was removed in this PR: https://github.com/adeira/universe/pull/2080. I had to disable it because it was miraculously failing on our CI and I was not able to figure out WTF. Let's see if it's better now after some dependency upgrades.