Shipit: fix `adeira-source-id` position when using `Co-authored-by` lines
Basically, we were always printing `adeira-source-id` line as a last one in the commit description. But turned out that `Co-authored-by` lines must be always the last ones for it to work correctly.

This change attempts to do exactly that: it separates the actual description and `Co-authored-by` lines when parsing the commit header (`parsePatchHeader`) and it constructs the commit description correctly in `addTrackingData` function.

Closes: https://github.com/adeira/universe/issues/2261