YCBO: refetch product revision when editing the product
We are sending product revision when editing the product to make sure the editation is sound and we won't lose any data. The revision changes on BE with every update so we have to refetch it every time to get the new value.

I also added some missing strings for the translation.