Remove support for non-standard `module` field in `package.json`
This commit removes support for non-standard `"module"` field in `package.json`. Purpose of this field was to implement support for ES modules in its early days. This used to be a common way how to do it in bundlers, see:

- Webpack: https://webpack.js.org/guides/author-libraries/#final-steps
- Rollup: https://github.com/rollup/rollup/wiki/pkg.module
- https://stackoverflow.com/questions/42708484/what-is-the-module-package-json-field-for

However, seems like the whole system around ES modules evolved a bit differently (top level field `"type": "module"` + field `"exports"`), see:

- https://nodejs.org/api/packages.html
- https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c
- https://babeljs.io/blog/2021/02/22/7.13.0#automatic-es-modules-in-babelruntime-12632httpsgithubcombabelbabelpull12632

This change is essentially a cleanup of our quirky and non-standard solution in order to be able to implement this modern approach.