Fetch: do not retry/timeout upload requests
I got into a situation where I wanted to upload something via GraphQL (POST + FormData) and it took too long to respond. It's very unfortunate when our `fetch` tries to retry in this situation because it basically means we are sending a request to "create something" multiple times. POST request in general are problematic for this reason, however, they are common in GraphQL world and I am not sure how to distinguish queries from mutations.

This commit changes the behavior a little bit (while keeping the old behavior intact): it does not retry when the request contains `FormData` payload.