Relay example: remove CI in exported repo
I would like to remove these files. Relay example keeps failing regularly because the CI configuration gets easily out of sync and I do not have energy for keeping it up to date. Moreover, failed tests in the exported repo are not stopping deployments to Vercel. The source of truth is `adeira/universe` where all the checks are always being executed.