SX Design: make `Image` and `ProductCard` responsive
This change is needed to accommodate smaller screens where the size of the product card should be smaller. Both `Image` and `ProductCard` components are now rendering into the parent component with aspect ratio 1:1. This way the users of this library can resize these components however they want. Aspect ratio 1:1 is currently hardcoded but can, of course, be changed when there is a need for it.