Colors: add new function `convertToRGBTriplet`
This commit adds new experimental function `convertToRGBTriplet` that takes any valid CSS/HTML color notation and converts it into RGB triplet that we use for internal calculations of accessibility, brightness and so on.

Use-case: I am building a `Text` component that renders accessible color based on the background. Internally, this component needs to access computed styles (background color) and browsers are returning the color in `rgb(…)` and `rgba(…)` formats (any maybe other formats as well?) and it's challenging to do these conversions in user code.

During the CSS Colors editations I discovered some limitations (most notably the triplet is not a great friend with alpha channel) but I will address these limitations later.