Colors: improve RGB(A) and HSL(A) detection
The goal of this change is to follow specification and how browsers are actually understanding colors. Basically, all previously valid colors are still valid, but the check is no longer benevolent. I verified all invalid cases in the browser to make sure they also do not accept these invalid colors.