Universe: remove Bazel files
The reasoning is fairly simple: nobody is actively working on
integrating with Bazel and there are no plans/capacity to do so. It
might be because Bazel is quite complex, and it's hard to set up it in a
way we need it to be. Moreover, it's blocking other efforts like this
one for example: https://github.com/adeira/universe/pull/2154 (Yarn 2).

Note: this doesn't mean that there is no need for something like Bazel.
However, it doesn't have to be Bazel (preferably something simpler that
does the job well enough).