Relay example: improve main navigation styles
The main goal of this change was to use `Text` component for the
navigation so we don't have to deal with the links color. This required
some small changes in `Text` component types as well.

Additionally, I slightly update `Link` styles so the navigation (or any
links, really) looks even better.