SX Design: improve `Tabs` accessibility
Previously, it was not possible to navigate the tabs via `Tab` and
`Tab+Shift` keys. So I changed the internal implementation to use HTML
`button` instead which enables it by default.

The accessibility is still not perfect. Specifically, we should be able
to change the tabs only by arrow keys (and not tabs). But that's for
later.