Relay: replace all loggers with one simple `RelayLogger`
This is a preparation for Relay 12.0.0 upgrade. I've found that I often
switch to use the debug logger because it provides much more useful
information when working with Relay Hooks. So I would like to prioritize
this logger and remove all the other loggers. We should make this one
work well for the new way of writing Relay applications.

Consequently, this makes it much easier to upgrade to Relay 12.0.0.