SX Design: apply word break to `ProductCard` component title
We have an adaptive font-size of product card title (when the title is
too long we change the font size). However, there are still some edge
cases when the product card can get too small (due to responsibility
in CSS grid). In this case we should apply word break so it still looks
nice and the title is not overflowing.