Dependabot: remove `automerge` label
I want to remote the `automerge` label because I want to remove the
requirement for 1 approval before merging. Let's be honest, I am the
only one working with this repo. 🤓

I think that if we'd keep this label then Kodiak would be merging all
the PRs directly without having an opportunity to review it first.