Shipit: correctly propagate PATH to `ShellCommand`
This change started with investigating why `src/monorepo-shipit/src/__tests__/RepoGit.findLastSourceCommit.test.js` fails on some machines. Here is an example of such failure:

```
  ● can find last source commit

    expect(received).toBe(expected) // Object.is equality

    Expected: "636abcdef012345840"
    Received: null

      27 |   const description = addTrackingData(new Changeset().withID(fakeCommitID)).getDescription();
      28 |   const repo = createGITRepoWithCommit(description);
    > 29 |   expect(repo.findLastSourceCommit(new Set())).toBe(fakeCommitID);
         |                                                ^
      30 | });
      31 |
      32 | it('can find last source commit with Co-authored-by', () => {

      at Object.toBe (src/__tests__/RepoGit.findLastSourceCommit.test.js:29:48)
```

Interestingly, this doesn't happen on CI. It happens only on some macOS machines (but doesn't fail using `macos-latest` on CI either). I've discovered that this is a problem of Git after reading the 2.40.0 changelog:

> Newer regex library macOS stopped enabling GNU-like enhanced BRE, where '\(A\|B\)' works as alternation, unless explicitly asked with the REG_ENHANCED flag.  "git grep" now can be compiled to do so, to retain the old behaviour.

Source: https://raw.githubusercontent.com/git/git/master/Documentation/RelNotes/2.40.0.txt

This lead me to the fix - install the latest Git version instead of using the broken one on macOS:

```
brew install git
```

This still however didn't work becuase Shipit was incorrectly working with PATH. Internally the Homebrew version was still being ignored.

The fix was to update the usage of `ShellCommand` because apparently when overwriting the envs, `ShellCommand` defaults to ignoring the PATH variable and therefore the incorrect Git version is used.

This change should fix the problem.