Shipit: improve CLI interface
The previous solution was good enough when the tool was used internally but now I see the need to have a better CLI interface. This not only provides better out-of-the box validations but also supports `--help` option.

I've also improved the Importit interface to support `--pull-request` option (instead of repo URL and PR number). I believe it's more natural to just copy the PR URL and use it.