CI: add new `rust-test-results` job
The purpose of this job is to wait for the completion of the `rust-test` jobs (matrix) and evaluate whether the run was successful or not. This is useful when configuring the required jobs in GitHub UI (we can require just this one job instead of all the jobs in the test matrix).